Update generate_toc_for_readme.sh to format Table of Contents
Updated the shell script generate_toc_for_readme.sh to clear the existing TOC.md file and generate a new table of contents in this file. This is achieved by obtaining the directory structure and writing it into TOC.md. A more detailed listing of files can now be found in TOC.md rather than in README.md.